requirements.txt
absl-py==2.3.0
astunparse==1.6.3
attrs==25.3.0
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
coremltools==8.3.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==2.15.0
scikit-learn==1.5.1
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes~=0.2.0
mpmath==1.3.0
namex==0.1.0
numpy>=1.23.5,<2.0.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
protobuf==3.20.3
pyaml==25.5.0
Pygments==2.19.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sympy==1.14.0
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorflow-macos==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt>=1.11.0,<1.15
librosa==0.10.1
